Changed peak decorators to optional features. Moved these to a separate file. Made family_closure more general. Moved ADT_fc functions to a utils file that can be removed in the future.
